remove model_js and place model_js elements into app_js, create test user and read test user
