--- data/models.js
@@ -1,4 +1,0 @@
-var mongoose = require('mongoose');
-var userSchema = require('./schema').userSchema;
-
-var User = mongoose.model('User', userSchema);
