--- data/routes.js
@@ -1,25 +1,0 @@
-/*var mongoose = require('mongoose');
-var schema = mongoose.Schema;
-var express = require('express');
-var app = express();
-var bodyParser = require('body-parser');
-
-
-app.use(bodyParser.urlencoded({
-    extended: true
-}));
-app.use(bodyParser.json());
-
-
-app.get('/', function(req, res) {
-    res.json({
-        message: 'This is the main page!',
-    });
-});
-
-app.listen(8080, function() {
-    console.log('Listening on port 8080');
-});
-
-exports.app = app;
-*/
